Move browse to submodule
